requirements.txt
asgiref==3.8.1
diff-match-patch==20241021
dj-database-url==2.3.0
Django==5.1.5
django-import-export==4.3.5
django-taggit==6.1.0
pillow==11.1.0
psycopg2==2.9.10
python-dotenv==1.0.1
sqlparse==0.5.3
tablib==3.8.0
typing_extensions==4.12.2
